Make sure the received field is a float

+++ BrianJacobs/Scheduler/database/migrations/2014_07_12_180544_create_user_credits.php
@@ -29,7 +29,7 @@ class CreateUserCredits extends Migration {
 
 		Schema::table('users_appointments', function(Blueprint $table)
 		{
-			$table->integer('received')->default(0)->after('amount');
+			$table->float('received')->default(0)->after('amount');
 		});
 	}
